Upload of Mobirise source
This is the source for Mobirise.  It can be copied to a mac/PC for future web development.   
Source includes a folder called unused assets which has some images that might be useful in the future.
It includes a folder called additional-pages that has 2 files that should always be kept in the published website root folder for Android play store compliance